requirements.txt
paho-mqtt==1.6.1
requests==2.28.1 
bleak==0.18.1